requirements.txt
streamlit 
google-generativeai 
sentence-transformers 
scikit-learn 
PyPDF2 
python-docx 
numpy